[NOID] Fixes #3831: Improve procedure errors in case of missing dependency (#3837)
