+++ full/src/main/java/apoc/data/email/ExtractEmail.java
@@ -18,38 +18,30 @@
  */
 package apoc.data.email;
 
+import static apoc.data.email.ExtractEmailHandler.extractEmail;
+
 import apoc.Extended;
-import apoc.util.Util;
+import apoc.util.MissingDependencyException;
 import java.util.Map;
-import javax.mail.internet.*;
 import org.neo4j.procedure.Description;
 import org.neo4j.procedure.Name;
 import org.neo4j.procedure.UserFunction;
 
 @Extended
 public class ExtractEmail {
+    public static final String EMAIL_MISSING_DEPS_ERROR =
+            "Cannot find the needed jar into the plugins folder in order to use . \n"
+                    + "Please put the apoc-email-dependencies-5.x.x-all.jar into plugin folder.\n"
+                    + "See the documentation: https://neo4j.com/labs/apoc/5/overview/apoc.data/apoc.data.email/#_install_dependencies";
 
     @UserFunction("apoc.data.email")
     @Description(
             "apoc.data.email('email_address') as {personal,user,domain} - extract the personal name, user and domain as a map")
     public Map<String, String> email(final @Name("email_address") String value) {
-        if (value == null || value.indexOf('@') == -1) {
-            return null;
-        }
         try {
-            InternetAddress addr = new InternetAddress(value);
-            String rawAddr = addr.getAddress();
-            int idx = rawAddr.indexOf('@');
-
-            return (Map) Util.map(
-                    "personal",
-                    addr.getPersonal(),
-                    "user",
-                    rawAddr.substring(0, idx),
-                    "domain",
-                    rawAddr.substring(idx + 1));
-        } catch (AddressException adr) {
-            return null;
+            return extractEmail(value);
+        } catch (NoClassDefFoundError e) {
+            throw new MissingDependencyException(EMAIL_MISSING_DEPS_ERROR);
         }
     }
 }
