+++ full/src/main/java/apoc/load/LoadHtml.java
@@ -53,6 +53,9 @@ public class LoadHtml {
 
     // public for test purpose
     public static final String KEY_ERROR = "errorList";
+    public static final String SELENIUM_MISSING_DEPS_ERROR = "Cannot find the Selenium client jar.\n"
+            + "Please put the apoc-selenium-dependencies-5.x.x-all.jar into plugin folder.\n"
+            + "See the documentation: https://neo4j.com/labs/apoc/5/overview/apoc.load/apoc.load.html/#selenium-dependencies";
 
     @Context
     public GraphDatabaseService db;
@@ -208,9 +211,7 @@ public class LoadHtml {
         try {
             return action.get();
         } catch (NoClassDefFoundError e) {
-            throw new MissingDependencyException(
-                    "Cannot find jars into the plugins folder.\n"
-                            + "See the documentation: https://neo4j.com/labs/apoc/4.1/overview/apoc.load/apoc.load.html/#selenium-depencencies");
+            throw new MissingDependencyException(SELENIUM_MISSING_DEPS_ERROR);
         }
     }
 }
